Fix and align unit tests for contributions from @glureau
