+++ README.md
@@ -1,7 +1,7 @@
 # Kotlin Fit Converter
 
 [![Kotlin](https://img.shields.io/badge/kotlin-2.0.0-blue.svg)](https://kotlinlang.org/)
-[![Latest Release](https://img.shields.io/badge/0.4.13-alpha-red)](https://github.com/example/garmin-fit-converter/releases)
+[![Latest Release](https://img.shields.io/badge/0.4.14-alpha-red)](https://github.com/example/garmin-fit-converter/releases)
 [![Gradle Build & Test](https://github.com/james-millner/kotlin-fit-converter/actions/workflows/gradle.yml/badge.svg)](https://github.com/james-millner/kotlin-fit-converter/actions/workflows/gradle.yml)
 
 The Kotlin Fit Converter Library is a Kotlin-based utility that allows users to convert Garmin FIT files into different formats.
